refactor: move algorithm classes to dedicated modules
