+++ TODO.md
@@ -45,6 +45,9 @@
 20c2.2g Convert delete_planned_bulk endpoint to async. [complete]
 20c2.2h Convert reorder_planned endpoint to async. [complete]
 20c2.2i Convert duplicate_planned_workout endpoint to async. [pending]
+20c2.2i.1 Asyncify `planner_service.duplicate_plan` method. [pending]
+20c2.2i.2 Update REST API `duplicate_planned_workout` endpoint to async. [pending]
+20c2.2i.3 Add tests for async duplicate plan endpoint. [pending]
 20c2.2j Convert auto_plan endpoint to async. [pending]
 20c2.2k Convert goal_plan endpoint to async. [pending]
 20c2.2l Convert use_planned_workout endpoint to async. [pending]
