+++ tests/test_longterm_usage.py
@@ -7,7 +7,7 @@ from fastapi.testclient import TestClient
 import sys
 sys.path.append(os.path.dirname(os.path.dirname(__file__)))
 from rest_api import GymAPI
-from tools import MathTools
+from algorithms.math_tools import MathTools
 
 
 class LongTermUsageTestCase(unittest.TestCase):
